Changes signature of ResourceProviderRegister.register_*(), adds load().
register_callable() and register_instance() now take `resource_name` then `provider`. Formerly it was vice-versa.

The load() method takes another ResourceProviderRegister instance and adds all of its registered providers to this one. By default, if the other register has providers for any resource names this register already has, an exception will be raised. If the `allow_overrides` arg is given as True, then instead of this behaviour, providers registered in the other register will _override_ the providers in this register for the same resource name.

The ResourceProviderRegister.__init__() method also now takes an optional `namespace` argument. Currently this does nothing, and is just provided as provisional interface compatibility.

In future, the ResourceProviderRegister and other fang classes will be upgraded to handle namespaces, and any non-absolute resource names will likely be considered relative to the specified namespace. The exact details of this still need to be finalised and implemented.