Expands DESCRIPTION.rst by re-working some of README.rst.
This re-works some of the text from README.rst, but removes some of the more "rationale"-type prose. I think this is less appropriate for the DESCRIPTION.rst, which (as I understand) is mainly used for the longer package description on PyPI. Since this description will mostly be seen those searching on PyPI or with pip, the prose should be drier and informative, rather than the conversation, "making the case" prose of the README.rst.