Adds unit tests for DependencyRegister.register()
This also includes a new test fixture in tests/unit/conftest.py: mock_dep_reg, which creates a mock instance of DependencyRegister.

This is used to ensure that DependencyRegister methods under tests only access other attributes and methods of the DependencyRegister instance in expected ways. We do this by calling our method under test as `DependencyRegister.our_method(mock_dep_reg, args)`. In this way, we are passing a mock DependencyRegister instance as `self` for the method.