Tests: Changes unit tests of DependencyRegister to use a mock instance as self when appropriate.
Particularly, the tests for DependencyRegister._register_dependent() and DependencyRegister._register_resource_dependency() now use the strategy of calling the method-under-test on the class and passing a mock instance in as "self".

For example, to test _register_dependent(), instead of using:

   DependencyRegister_instance._register_dependent(fake_dependent, fake_resource_name)

...we now use:

   DependencyRegister._register_dependent(
           mock_DependencyRegister_instance, fake_dependent, fake_resource_name)

mock_DependencyRegister_instance is used as "self" within the method call, so we can check for, stub and mock any calls to other methods on self.