Tests: Adds @argument_line() decorator and related functions to BDD tests.
This allows registered strings and parsing patterns for argument lines with test steps, similarly to how "Given", "When" and "Then" steps as a whole are registered.

This is intended to replace the current ARG_LINES system in the BDD tests, with the improvement that the new system allows pattern matching with the pytest_bdd parsers, for parameterised argument fixtures.

This has primarily been added to support a scenario with the step:

    Then the method _register_dependent should be called with:
        the return value of method _unwrap_dependent
        a fake resource name

In the above step, I wanted the first argument line to use parsing and pattern matching, eg "the return value of method {method_name}",
 with the "fixture" function for the argument value to support any named method, rather than hard-coding a fixed string just for "_unwrap_dependent".