Tests: Adds feature DependencyRegister.query_resources with one scenario.
This feature and scenario correspond to the non-BDD test class Test_DependencyRegister_query_resources and its method test__giving_dependent__should_call_self__unwrap_dependent().

This also includes code for supporting a new Given step, "Given I am ignoring all exceptions during the method call". This wraps the call of the method under test in a try/except. This is done because this particular scenario is not concerned with testing exception raising - this is checked for in another test.