Tests: Adds new scenario for feature DependencyRegister.query_resources and new "defers_when_steps" system.
The new system allows decorating a Then step with "@defers_when_steps". This causes the When steps to be removed from the normal flow, and instead passed as a function to the Then steps as a new argument "deferred_when_steps".

The allows us to easily express the "Then the exception {exception_name} should be raised" step as simply wrapping the deferred_when_steps() call in a "with pytest.raises(...):" block.

Unfortunately, the implementation of this is kinda ugly, and involves knowing a bit too much about pytest_bdd's internal details.The implementation uses some pytest hooks for pytest_bdd(), which are fine, but also modifies the Scenario object and accesses some underscore-prefixed functions on the pytest_bdd.scenario module.